+++ pkg/appsec/appsec.go
@@ -10,6 +10,12 @@ import (
 var (
 	// ErrStructValidation is returned when given struct validation failed
 	ErrStructValidation = errors.New("struct validation")
+
+	// ErrRequestCreation is returned when creating an HTTP request failed
+	ErrRequestCreation = errors.New("HTTP request failure")
+
+	// ErrAPICallFailure is returned when an Appsec OpenAPI call failed
+	ErrAPICallFailure = errors.New("API call failure")
 )
 
 type (
@@ -19,6 +25,7 @@ type (
 		AdvancedSettingsAttackPayloadLogging
 		AdvancedSettingsEvasivePathMatch
 		AdvancedSettingsLogging
+		AdvancedSettingsPIILearning
 		AdvancedSettingsPragma
 		AdvancedSettingsPrefetch
 		AdvancedSettingsRequestBody