+++ pkg/appsec/mocks.go
@@ -360,6 +360,14 @@ func (m *Mock) UpdateAdvancedSettingsEvasivePathMatch(ctx context.Context, req U
 	return args.Get(0).(*UpdateAdvancedSettingsEvasivePathMatchResponse), args.Error(1)
 }
 
+func (m *Mock) UpdateAdvancedSettingsPIILearning(ctx context.Context, req UpdateAdvancedSettingsPIILearningRequest) (*AdvancedSettingsPIILearningResponse, error) {
+	args := m.Called(ctx, req)
+	if args.Get(0) == nil {
+		return nil, args.Error(1)
+	}
+	return args.Get(0).(*AdvancedSettingsPIILearningResponse), args.Error(1)
+}
+
 func (m *Mock) UpdateAdvancedSettingsRequestBody(ctx context.Context, req UpdateAdvancedSettingsRequestBodyRequest) (*UpdateAdvancedSettingsRequestBodyResponse, error) {
 	args := m.Called(ctx, req)
 	if args.Get(0) == nil {
@@ -1203,6 +1211,14 @@ func (m *Mock) GetAdvancedSettingsEvasivePathMatch(ctx context.Context, req GetA
 		return nil, args.Error(1)
 	}
 	return args.Get(0).(*GetAdvancedSettingsEvasivePathMatchResponse), args.Error(1)
+}
+
+func (m *Mock) GetAdvancedSettingsPIILearning(ctx context.Context, req GetAdvancedSettingsPIILearningRequest) (*AdvancedSettingsPIILearningResponse, error) {
+	args := m.Called(ctx, req)
+	if args.Get(0) == nil {
+		return nil, args.Error(1)
+	}
+	return args.Get(0).(*AdvancedSettingsPIILearningResponse), args.Error(1)
 }
 
 func (m *Mock) GetAdvancedSettingsRequestBody(ctx context.Context, req GetAdvancedSettingsRequestBodyRequest) (*GetAdvancedSettingsRequestBodyResponse, error) {